Move sub-classes of "MdiClient", "MdiWindowDialog", "RichTextBox" and "ScrollableControl" classes to their own files
